+++ frontend/src/components/FocusOverlay.js
