+++ frontend/src/components/InfoBox.js
@@ -4,8 +4,8 @@ import { Button, Paper, Typography, IconButton, Popper } from '@material-ui/core
 import { InfoOutlined as InfoIcon } from '@material-ui/icons'
 
 import { useFocusOverlay } from './FocusOverlay'
-import userGuide from '../../static/userGuide'
-import { setProgress, getUser } from '../../lib/userProgress'
+import userGuide from '../static/userGuide'
+import { setProgress, getUser } from '../lib/userProgress'
 
 const useStyles = makeStyles(theme => ({
   root: {
