+++ frontend/src/views/graph/colors.json
