+++ frontend/src/views/graph/generateColors.js
