+++ frontend/src/views/mapper/GuidedCourseContainer.js
@@ -6,7 +6,7 @@ import Masonry from './Masonry'
 import useCreateConceptDialog from '../../dialogs/concept/useCreateConceptDialog'
 import useEditConceptDialog from '../../dialogs/concept/useEditConceptDialog'
 import useEditCourseDialog from '../../dialogs/course/useEditCourseDialog'
-import { useInfoBox } from '../common/InfoBox'
+import { useInfoBox } from '../../components/InfoBox'
 
 const GuidedCourseContainer = ({
   courseTrayOpen,
