+++ frontend/src/views/mapper/Masonry.js
