+++ frontend/src/views/project/CloneView.js
