+++ frontend/src/views/resource/Resource.js
